메세지 수, 리액션 데이터 불러오기
